pyproject.toml
[tool.poetry]
name = "sz-bot"
version = "2.1.0"
description = "Private bot for Discord's 'Study Zone Server'"
authors = ["Alimus SIfar <alimus.sifar@g.bracu.ac.bd>"]

[tool.poetry.dependencies]
python = "^3.8"
replit = "^3.2.4"
colorama = "^0.4.5"
python-dotenv = "^0.20.0"
Flask = "^2.1.2"
"discord.py" = {extras = ["voice"], version = "^2.0.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=1.1.0"]
